requirements.txt
Flask==2.3.3
langchain_google_genai==0.0.4
gunicorn==20.1.0
python-dotenv==1.0.0